split main in classes
